Create AlignAndPickUpSkystone.java
Can detect skystones and retrieve position info. However, it does not indicate when it goes to the next state (or it doesn't go to the next state at all). Going to next state also disallows the robot from detecting skystones (or at least from updating telemetry). If a timer is added to make the robot wait before going to the next state, vuforia works until the timer runs out, then the robot does stops updating telemetry (so I can't tell if it is actually going to the next state).