Structural Cahnges to V2 and Incorperated Vuforia into V1
Encorperated new Vuforia Data class into StrafeToSkystone instead of StrafeToSkystoneV2 because V2 depends on ability for class to determine quarry configuration which is unable to be done for now. Instead, the vurofria class was encorperated into StrafeToSkystone which has a structure that can better match the capabilities of the vuforia class for now.