Update Two Skystone and Park Auto - may have broken it
The distance when driving back for the second skystone works, and the robot does less unnecessary strafing. However, the robot no longer grabs the second skystone; it drives in the opposite direction it's supposed to for adjusting to pick up the skystone. Previously, it did not adjust incorrectly, so this change may have to be reverted.