Update PathThreeTwoSkystonesAndPark.java
Figured out previous problem - the robot's detection of the second skystone's position was messed with since it started detecting before the robot stopped moving. Adding a wait before detecting the second skystone solved this issue.

The two skystone positions furthest from the wall work on both red and blue sides. Special cases for the position closest to the wall need to be added.